Split StartGame for lazy loading
